requirements.txt
matplotlib
pytest
google-generativeai
